requirements.txt
allure-behave==2.13.1
allure-python-commons==2.13.1
attrs==22.2.0
behave==1.2.6
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
requests==2.28.2
six==1.16.0
urllib3==1.26.15
